Welcome Blade Modifications. My local copy of the site displayed on the computer perfectly. The staging environment on vapor displayed a 500 message. I took a look at the AWS logs and noticed an issue within the Hero section, specifically the navigation section. I removed the navigation section from the project temporarily and deployed to the staging environment. I can view the web page correctly.
I remove the Video placeholder image and replaced it the new company logo that I created. With images, you must save them in the public folder if you want them to dipslay in the staging environment.  You can use a AWS URL but I wanted to go the short route.

I need to remove the database and the migrate deployment hook from staging as this site does not have a back end.